requirements.txt
# Core
pandas
numpy
scikit-learn
lightgbm
xgboost
joblib
kagglehub

# Configuration
pyyaml

# Ingestion
requests

# Tracking
mlflow

# Web API
fastapi
uvicorn

# Workflow & Monitoring
prefect
evidently

# Dev & infra
python-dotenv

#tests
pytest

# code formatting pep8
black